+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Support auto scaling for math block ([#30](https://github.com/marp-team/marp-core/pull/30))
+
 ## v0.0.6 - 2018-09-06
 
 ### Fixed